Improve update-driver script for MLNX 4.2 driver

+++ drivers/4.2/orig/libmlx4-41mlnx1/update-driver.sh
@@ -1,3 +1,9 @@
-#sudo make
-sudo cp $(dirname $0)/src/.libs/*-rdmav2.so /usr/lib/libibverbs/
+# make
+if [ -f /etc/fedora-release ]; then
+  sudo cp $(dirname $0)/src/.libs/libmlx4-rdmav2.so /usr/lib64/libmlx4-rdmav2.so
+fi
+
+if [ -f /etc/lsb-release ]; then
+  sudo cp $(dirname $0)/src/.libs/*-rdmav2.so /usr/lib/libibverbs/
+fi
